requirements-gui.txt
PySide6>=6.6.0
PySide6-Addons>=6.6.0
PySide6-Essentials>=6.6.0
PyInstaller>=6.0.0
Pillow>=10.0.0
requests>=2.28.0
